# specs/1-todo-web-app/contracts/api-contract.yaml
openapi: 3.0.0
info:
  title: Todo Evolution Hackathon API
  description: API for Phase 2 Full-Stack Todo Web Application
  version: 1.0.0
  contact:
    name: Todo Evolution Team
    email: dev@todo-evolution.hackathon

servers:
  - url: http://localhost:8000/api
    description: Development server
  - url: https://api.todo-evolution.hackathon/api
    description: Production server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Task:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - priority
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the task
        user_id:
          type: string
          format: uuid
          description: Reference to the user who owns the task
        title:
          type: string
          minLength: 1
          maxLength: 100
          description: Task title
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: Extended description of the task
        completed:
          type: boolean
          description: Whether the task is completed
          default: false
        priority:
          type: string
          enum: [high, medium, low]
          default: medium
          description: Priority level of the task
        tags:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 50
          maxItems: 10
          description: Array of tag strings associated with the task
          default: []
        due_date:
          type: string
          format: date-time
          nullable: true
          description: Optional deadline for the task
        created_at:
          type: string
          format: date-time
          description: When the task was created
        updated_at:
          type: string
          format: date-time
          description: When the task was last updated

    TaskInput:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
          description: Task title
        description:
          type: string
          maxLength: 1000
          description: Extended description of the task
        priority:
          type: string
          enum: [high, medium, low]
          default: medium
          description: Priority level of the task
        tags:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 50
          maxItems: 10
          description: Array of tag strings associated with the task
        due_date:
          type: string
          format: date-time
          nullable: true
          description: Optional deadline for the task
        completed:
          type: boolean
          description: Whether the task is completed
          default: false

    TaskResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/Task'

    TaskListResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        total:
          type: integer
          description: Total number of tasks matching the query
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Number of tasks per page

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details

    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          description: Success message

  parameters:
    taskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: The ID of the task

  responses:
    UnauthorizedError:
      description: Unauthorized - Invalid or missing JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

security:
  - bearerAuth: []

paths:
  /tasks:
    get:
      summary: Get user's tasks
      description: Retrieve a list of tasks for the authenticated user with optional filtering and sorting
      parameters:
        - name: filter_status
          in: query
          schema:
            type: string
            enum: [completed, incomplete]
          description: Filter by completion status
        - name: filter_priority
          in: query
          schema:
            type: string
            enum: [high, medium, low]
          description: Filter by priority level
        - name: filter_tag
          in: query
          schema:
            type: string
          description: Filter by tag
        - name: search
          in: query
          schema:
            type: string
          description: Search keyword in title/description
        - name: sort
          in: query
          schema:
            type: string
            enum: [due_date, priority, title, created_at, updated_at]
          description: Sort field
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Sort order (ascending or descending)
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number for pagination
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Number of tasks per page
      responses:
        '200':
          description: List of tasks for the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      summary: Create a new task
      description: Create a new task for the authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskInput'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Bad request - Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /tasks/{task_id}:
    get:
      summary: Get a specific task
      description: Retrieve a specific task by ID for the authenticated user
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      summary: Update a task
      description: Update an existing task for the authenticated user
      parameters:
        - $ref: '#/components/parameters/taskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskInput'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Bad request - Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      summary: Delete a task
      description: Delete a specific task for the authenticated user
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /tasks/{task_id}/complete:
    patch:
      summary: Toggle task completion status
      description: Mark a task as complete or incomplete for the authenticated user
      parameters:
        - $ref: '#/components/parameters/taskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                completed:
                  type: boolean
                  description: Whether the task is completed
      responses:
        '200':
          description: Task completion status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Bad request - Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'